+++ frontend/app/layout.tsx
@@ -1,7 +1,9 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
+
 import Providers from "./Providers";
+import AppShell from "./AppShell";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -14,22 +16,25 @@ const geistMono = Geist_Mono({
 });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "FinSight",
+  description: "Stock analytics and learning platform",
 };
 
 export default function RootLayout({
   children,
-}: Readonly<{
+}: {
   children: React.ReactNode;
-}>) {
+}) {
   return (
     <html lang="en">
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
+        style={{ display: "flex", flexDirection: "column", minHeight: "100vh" }}
       >
-        {/* ✅ Wrap the entire app with Providers */}
-        <Providers>{children}</Providers>
+        <Providers>
+          {/* Client-side UI */}
+          <AppShell>{children}</AppShell>
+        </Providers>
       </body>
     </html>
   );
