+++ frontend/app/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import React, { useState } from "react";
+import React from "react";
 import {
   Box,
   Container,
@@ -9,12 +9,7 @@ import {
   CardContent,
   Divider,
   Grid,
-  Dialog,
 } from "@mui/material";
-import Header from "@/components/Header";
-import Footer from "@/components/Footer";
-import LoginForm from "@/components/LoginForm";
-import SignupForm from "@/components/SignupForm";
 
 const stocks = [
   {
@@ -56,148 +51,74 @@ const stocks = [
 ];
 
 export default function HomePage() {
-  const [openLogin, setOpenLogin] = useState(false);
-  const [openSignup, setOpenSignup] = useState(false);
-
-  const handleOpenLogin = () => {
-    setOpenLogin(true);
-    setOpenSignup(false);
-  };
-  const handleOpenSignup = () => {
-    setOpenSignup(true);
-    setOpenLogin(false);
-  };
-
   return (
-    <Box sx={{ width: "100%", overflowX: "hidden" }}>
-      <Header onLoginClick={handleOpenLogin} onSignupClick={handleOpenSignup} />
+    <Box sx={{ width: "100%", px: { xs: 2, md: 4 }, py: 4 }}>
+      <Container maxWidth="lg">
+        <Typography variant="h4" fontWeight={700} gutterBottom>
+          Welcome back, Investor 👋
+        </Typography>
 
-      {/* Main content */}
-      <Box
-        component="main"
-        sx={{
-          px: { xs: 2, md: 4 },
-          py: 4,
-          transition: "0.3s",
-          filter: openLogin || openSignup ? "blur(6px)" : "none",
-          bgcolor: "background.paper",
-        }}
-      >
-        <Container maxWidth="lg">
-          <Typography variant="h4" fontWeight={700} gutterBottom>
-            Welcome back, Investor 👋
-          </Typography>
-          <Typography variant="body1" color="text.secondary" mb={6}>
-            Research and go for it.
-          </Typography>
+        <Typography variant="body1" color="text.secondary" mb={6}>
+          Research and go for it.
+        </Typography>
 
-          <Grid container spacing={3}>
-            {stocks.map((s) => (
-              <Grid item xs={12} sm={6} md={3} key={s.symbol}>
-                <Card
-                  sx={{
-                    borderRadius: 3,
-                    boxShadow: 2,
-                    "&:hover": { boxShadow: 6 },
-                    transition: "0.3s",
-                    bgcolor: "background.paper",
-                  }}
-                >
-                  <CardContent>
-                    <Typography
-                      variant="subtitle1"
-                      fontWeight={700}
-                      color="primary"
-                    >
-                      {s.symbol}
-                    </Typography>
-                    <Typography variant="body2" color="text.secondary">
-                      {s.name}
-                    </Typography>
-                    <Typography variant="h5" fontWeight={700} mt={2}>
-                      ${s.price}
-                    </Typography>
-                    <Typography variant="body2" sx={{ color: s.color }}>
-                      {s.delta}
-                    </Typography>
-                    <Divider sx={{ my: 2 }} />
-                    <Box display="flex" justifyContent="space-between">
-                      <Box>
-                        <Typography variant="caption" color="text.secondary">
-                          7-Day Prediction
-                        </Typography>
-                        <Typography variant="body2" color="primary">
-                          {s.prediction}
-                        </Typography>
-                      </Box>
-                      <Box textAlign="right">
-                        <Typography variant="caption" color="text.secondary">
-                          Confidence
-                        </Typography>
-                        <Typography variant="body2">{s.confidence}</Typography>
-                      </Box>
+        <Grid container spacing={3}>
+          {stocks.map((s) => (
+            <Grid item xs={12} sm={6} md={3} key={s.symbol}>
+              <Card
+                sx={{
+                  borderRadius: 3,
+                  boxShadow: 2,
+                  "&:hover": { boxShadow: 6 },
+                  transition: "0.3s",
+                }}
+              >
+                <CardContent>
+                  <Typography
+                    variant="subtitle1"
+                    fontWeight={700}
+                    color="primary"
+                  >
+                    {s.symbol}
+                  </Typography>
+
+                  <Typography variant="body2" color="text.secondary">
+                    {s.name}
+                  </Typography>
+
+                  <Typography variant="h5" fontWeight={700} mt={2}>
+                    ${s.price}
+                  </Typography>
+
+                  <Typography variant="body2" sx={{ color: s.color }}>
+                    {s.delta}
+                  </Typography>
+
+                  <Divider sx={{ my: 2 }} />
+
+                  <Box display="flex" justifyContent="space-between">
+                    <Box>
+                      <Typography variant="caption" color="text.secondary">
+                        7-Day Prediction
+                      </Typography>
+                      <Typography variant="body2" color="primary">
+                        {s.prediction}
+                      </Typography>
                     </Box>
-                  </CardContent>
-                </Card>
-              </Grid>
-            ))}
-          </Grid>
-        </Container>
-      </Box>
 
-      <Footer />
-
-      {/* Login Modal */}
-      <Dialog
-        open={openLogin}
-        onClose={() => setOpenLogin(false)}
-        PaperProps={{
-          sx: {
-            borderRadius: 3,
-            width: "400px",
-            maxWidth: "90%",
-            padding: 2,
-            margin: "auto",
-          },
-        }}
-        BackdropProps={{
-          sx: {
-            backgroundColor: "rgba(0,0,0,0.4)",
-            backdropFilter: "blur(6px)",
-          },
-        }}
-      >
-        <LoginForm
-          closeModal={() => setOpenLogin(false)}
-          switchToSignup={handleOpenSignup}
-        />
-      </Dialog>
-
-      {/* Signup Modal */}
-      <Dialog
-        open={openSignup}
-        onClose={() => setOpenSignup(false)}
-        PaperProps={{
-          sx: {
-            borderRadius: 3,
-            width: "400px",
-            maxWidth: "90%",
-            padding: 2,
-            margin: "auto",
-          },
-        }}
-        BackdropProps={{
-          sx: {
-            backgroundColor: "rgba(0,0,0,0.4)",
-            backdropFilter: "blur(6px)",
-          },
-        }}
-      >
-        <SignupForm
-          closeModal={() => setOpenSignup(false)}
-          switchToLogin={handleOpenLogin}
-        />
-      </Dialog>
+                    <Box textAlign="right">
+                      <Typography variant="caption" color="text.secondary">
+                        Confidence
+                      </Typography>
+                      <Typography variant="body2">{s.confidence}</Typography>
+                    </Box>
+                  </Box>
+                </CardContent>
+              </Card>
+            </Grid>
+          ))}
+        </Grid>
+      </Container>
     </Box>
   );
 }
