+++ frontend/package.json
@@ -21,6 +21,7 @@
     "d3-format": "^3.1.0",
     "d3-scale": "^4.0.2",
     "d3-time-format": "^4.1.0",
+    "framer-motion": "^12.23.25",
     "lucide-react": "^0.556.0",
     "next": "16.0.1",
     "react": "^18.2.0",
