+++ frontend/src/components/Header.tsx
@@ -64,7 +64,7 @@ export default function Header({ onLoginClick, onSignupClick }: HeaderProps) {
   const navLinks = [
     { label: "Home", href: "/" },
     { label: "News", href: "/news" },
-    { label: "About Us", href: "/about-us" },
+    { label: "About Us", href: "/aboutus" },
     { label: "Analysis", href: "/analysis" },
   ];
